backend/config/requirements-dev.txt
pytest
pytest-django
model-bakery
ruff
black